+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * Added tags @robgietema
 * Upgrade to React 16 @sneridagh
 * Upgrade to Webpack 4 @sneridagh
+* Renew login token when almost expired @robgietema
 
 ## 0.4.0 (2017-05-03)